+++ src/impl/index.ts
@@ -1,1 +1,4 @@
-export { TeamsBot } from "./TeamsBot";
+// Copyright (c) Wictor Wilén. All rights reserved.
+// Licensed under the MIT license.
+
+export { MessagingExtensionMiddleware } from "./MessagingExtensionMiddleware";